requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pillow==10.0.0
transformers==4.35.0
torch==2.1.0
python-multipart==0.0.6
